l'ajoutation d'une methode pour rechercher des parents dans FamilleController et l'ajoutation de folder famille dans folder admin

+++ routes/web.php
@@ -16,7 +16,7 @@ Route::middleware('auth')->group(function () {
         Route::view('/', 'admin.index')->name('index');
         Route::view('/publications', 'admin.messages')->name('messages');
         Route::view('/comptes', 'admin.accounts')->name('accounts');
-        Route::view('/familles', 'admin.families')->name('families');
+       Route::view('/familles', 'admin.familles.index')->name('admin.familles.index');
         Route::view('/classes', 'admin.classes')->name('classes');
         Route::view('/facture', 'admin.invoices')->name('invoices');
         Route::view('/notifications', 'admin.notifications')->name('notifications');
@@ -27,12 +27,12 @@ Route::get('/presence', function () {
     return view('presence.index');
 })->name('presence.index');
 
-Route::get('/familles', [FamilleController::class, 'index'])->name('familles.index');
-Route::get('/familles/create', [FamilleController::class, 'createView'])->name('familles.create');
-Route::get('/familles/{id}', [FamilleController::class, 'show'])->name('familles.show');
-Route::get('/familles/{id}/edit', [FamilleController::class, 'editView'])->name('familles.edit');
-Route::put('/familles/{id}', [FamilleController::class, 'update'])->name('familles.update');
-Route::delete('/familles/{id}', [FamilleController::class, 'delete'])->name('familles.delete');
+Route::get('/admin/familles', [FamilleController::class, 'index'])->name('admin.familles.index');
+Route::get('/familles/create', [FamilleController::class, 'createView'])->name('admin.familles.create');
+Route::get('/admin/familles/{id}', [FamilleController::class, 'show'])->name('admin.familles.show');
+Route::get('/familles/{id}/edit', [FamilleController::class, 'editView'])->name('admin.familles.edit');
+Route::put('/familles/{id}', [FamilleController::class, 'update'])->name('admin.familles.update');
+Route::delete('/familles/{id}', [FamilleController::class, 'delete'])->name('admin.familles.delete');
 
 
 Route::get('/presence/classes', [PresenceController::class, 'classes'])->name('presence.classes');
